#1 Denormalize data model: * delete SkillReport and UserReport * rename SkillPriorityReport  to UserSkillPriorityAggregationReport * correct java docs and swagger docs

+++ src/main/java/io/knowledgeassets/myskills/server/report/userskillpriorityreport/UserSkillPriorityReportResponse.java
@@ -10,5 +10,5 @@ import java.time.LocalDateTime;
 public class UserSkillPriorityReportResponse {
 	private String id;
 	private LocalDateTime date;
-	private Long skillCount;
+	private Integer skillCount;
 }
